refactor: reorganize retrieve code
